+++ .vscode/settings.json
@@ -1,14 +1,21 @@
 {
     "files.associations": {
         "__bit_reference": "cpp",
+        "__bits": "cpp",
+        "__config": "cpp",
+        "__debug": "cpp",
+        "__errc": "cpp",
         "__hash_table": "cpp",
         "__locale": "cpp",
+        "__mutex_base": "cpp",
         "__node_handle": "cpp",
         "__split_buffer": "cpp",
         "__threading_support": "cpp",
-        "__tree": "cpp",
+        "__tuple": "cpp",
         "__verbose_abort": "cpp",
         "array": "cpp",
+        "atomic": "cpp",
+        "bit": "cpp",
         "bitset": "cpp",
         "cctype": "cpp",
         "clocale": "cpp",
@@ -23,37 +30,43 @@
         "ctime": "cpp",
         "cwchar": "cpp",
         "cwctype": "cpp",
-        "deque": "cpp",
-        "execution": "cpp",
-        "memory": "cpp",
-        "fstream": "cpp",
+        "exception": "cpp",
         "initializer_list": "cpp",
-        "iomanip": "cpp",
         "ios": "cpp",
         "iosfwd": "cpp",
         "iostream": "cpp",
         "istream": "cpp",
         "limits": "cpp",
         "locale": "cpp",
+        "memory": "cpp",
         "mutex": "cpp",
         "new": "cpp",
         "optional": "cpp",
         "ostream": "cpp",
-        "print": "cpp",
-        "queue": "cpp",
         "ratio": "cpp",
-        "set": "cpp",
         "sstream": "cpp",
-        "stack": "cpp",
         "stdexcept": "cpp",
         "streambuf": "cpp",
         "string": "cpp",
         "string_view": "cpp",
+        "system_error": "cpp",
         "tuple": "cpp",
+        "type_traits": "cpp",
         "typeinfo": "cpp",
         "unordered_map": "cpp",
         "variant": "cpp",
         "vector": "cpp",
-        "algorithm": "cpp"
+        "__nullptr": "cpp",
+        "__string": "cpp",
+        "chrono": "cpp",
+        "compare": "cpp",
+        "concepts": "cpp",
+        "algorithm": "cpp",
+        "numeric": "cpp",
+        "random": "cpp",
+        "deque": "cpp",
+        "fstream": "cpp",
+        "iomanip": "cpp",
+        "stack": "cpp"
     }
 }
