+++ PA5/TournamentNode.cpp
@@ -1,17 +1,6 @@
-/*
-AUTHOR INFO
-Names: Joshua Vaysman, Divi Newton
-Student IDs: 2449656, 2440117
-Emails: vaysman@chapman.edu, dnewton@chapman.edu
-Class Section: CPSC-350-02
-Assignment: PA5 - Scare Games
-*/
-
 #include "TournamentNode.h"
 #include <iostream>
 
-// default constructor
-// sets all variables to NULL
 TournamentNode::TournamentNode() {
     m_data = Monster();
     m_winner = NULL;
@@ -19,8 +8,6 @@ TournamentNode::TournamentNode() {
     m_right = NULL;
 }
 
-// overloaded constructor
-// sets variables to NULL and sets data to a specified Monster
 TournamentNode::TournamentNode(Monster monster) {
     m_data = monster;
     m_winner = NULL;
@@ -28,23 +15,23 @@ TournamentNode::TournamentNode(Monster monster) {
     m_right = NULL;
 }
 
-// destructor
 TournamentNode::~TournamentNode() {
     
 }
 
-// fight method returns the loser based on two monsters fighting each other
-// contains all nessesary logic for fighting monsters in a node
+
+// seg faults if it is a by
+// don't think we need to worry about it as I assigned by's manually in TournamentTree
 Monster TournamentNode::fight() {
     Monster loser;
     std::cout << "Fighting: " << m_left->m_data.getName() << " vs " << m_right->m_data.getName() << std::endl;
-    if (m_left -> m_data.getScream() == m_right -> m_data.getScream()) { // if the screams are equal
+    if (m_left -> m_data.getScream() == m_right -> m_data.getScream()) {
         m_winner = m_left;
         setData(m_winner->m_data);
         loser = m_right->m_data;
         return loser;
     }
-    if (m_right -> m_hasData) { // if the right side has data
+    if (m_right -> m_hasData) {
         if (m_left->m_data.screamFight(m_right->m_data)) {
             m_winner = m_left;
             setData(m_winner->m_data);
@@ -53,7 +40,7 @@ Monster TournamentNode::fight() {
             m_winner = m_right;
             loser = m_left->m_data;
         }   
-    } else { // if right side has no data
+    } else {
         if (m_right == nullptr) {
             m_winner = m_left;
             setData(m_winner -> m_data);
@@ -64,12 +51,51 @@ Monster TournamentNode::fight() {
         loser = m_right -> m_data;
     }
     std::cout << "Loser: " << loser.getName() << std::endl;
-    return loser; // returns the looser
+    return loser;
 }
 
 
-// setData method sets the data of a node (the winner) to a specific monster
-// used from fight method to define the winner in a TournamentNode object
+
 void TournamentNode::setData(Monster monster) {
     m_data = monster;
 }
+
+// int main(int argc, char const *argv[])
+// {
+//     Monster meh("Mike", 4);
+//     Monster A("A", 1);
+//     Monster scarey("Sulley", 10);
+//     Monster B("B", 0);
+
+//     TournamentNode* root = new TournamentNode();
+//     TournamentNode* rl = new TournamentNode();
+//     TournamentNode* rr = new TournamentNode();
+
+
+
+//     TournamentNode* l1 = new TournamentNode(meh);
+//     TournamentNode* l2 = new TournamentNode(A);
+//     TournamentNode* r1 = new TournamentNode(scarey);
+//     TournamentNode* r2 = new TournamentNode(B);
+
+
+//     // switch this back to private after testing
+//     root -> m_left = rl;
+//     root -> m_right = rr;
+
+//     rl -> m_left = l1;
+//     rl -> m_right = l2;
+
+//     rr -> m_left = r1;
+//     rr -> m_right = r2;
+
+
+//     rl->fight();
+//     // rl->setData(rl -> m_winner -> m_data);
+//     rr->fight();
+//     // rr->setData(rr -> m_winner -> m_data);
+//     root -> fight();
+
+
+//     return 0;
+// }
