+++ PA5/TournamentTree.h
@@ -1,21 +1,6 @@
-/*
-AUTHOR INFO
-Names: Joshua Vaysman, Divi Newton
-Student IDs: 2449656, 2440117
-Emails: vaysman@chapman.edu, dnewton@chapman.edu
-Class Section: CPSC-350-02
-Assignment: PA5 - Scare Games
-*/
-
-#ifndef TOURNAMENTTREE_H
-#define TOURNAMENTTREE_H
-
 #include "TournamentNode.h"
 #include <fstream>
 
-// defines TournamentTree class
-// class includes public constructor and destructor, and other methods
-// class includes private m_bracket, m_reader, m_file, m_bracketSize, and nessesary helper methods
 class TournamentTree
 {
 private:
@@ -40,8 +25,4 @@ public:
     Monster finalWinner();
     void printLosers();
     int calculateMinHeight(int numLeaves);
-    void saveTreeAsDotHelper(TournamentNode* node, std::ofstream& file, int& nodeID);
-    void saveTreeAsDot(const std::string& filename, TournamentNode* rootNode);
 };
-
-#endif
